Moved CLI logic from interface to logic.
