+++ algorithm_tester/tester_logic.py
@@ -1,6 +1,9 @@
 import os
 import timeit
-from typing import Dict
+import click
+import time
+from typing import Dict, List
+from algorithm_tester.helpers import create_path
 from algorithm_tester.abstracts import TesterContext
 from algorithm_tester.plugins import plugins
 
@@ -15,6 +18,17 @@ def inner(_it, _timer{init}):
     return _t1 - _t0, ret_val
 """
 timeit.template = new_template
+
+def create_columns_description_file(algorithm: str, check_time: bool, output_dir: str):
+    column_descriptions = plugins.get_algorithm(name=algorithm).get_additional_columns(check_time)
+
+    with open(f'{output_dir}/column_description_{algorithm}.dat', "w") as f:
+        f.write(f'{" ".join(column_descriptions)}\n')
+
+def write_solution(output_file, parsed_data: Dict[str, object]):
+    # output_file.write(f"{solution.output_str()}\n")
+    output: str = f'{parsed_data["id"]} {parsed_data["item_count"]} {parsed_data["max_value"]} {parsed_data["algorithm_name"]} {parsed_data["elapsed_configs"]} | {" ".join(map(str, parsed_data["things"]))}'
+    output_file.write(f'{output}\n')
 
 def get_instance_file_results(datafile, algorithm: str, check_time: bool, time_retries: int, other_options: Dict[str, object] = None):
     data = datafile.readline()
@@ -52,3 +66,24 @@ def get_instance_file_results(datafile, algorithm: str, check_time: bool, time_r
         data = datafile.readline()
     
     print
+
+def run_algorithms_for_file(parser: str, algorithms: List[str], check_time: bool, time_retries: int, other_options: Dict[str, object], input_file, output_dir):
+    create_path(output_dir)
+
+    for algorithm in algorithms:
+        input_file.seek(0)
+        create_columns_description_file(algorithm, check_time, output_dir)
+
+        suffix = f"_{algorithm}"
+
+        output_file_name = input_file.name.split("/")[-1].replace(".dat", f'{suffix}.dat')
+        
+        it = get_instance_file_results(datafile=input_file, check_time=check_time, 
+            algorithm=algorithm, time_retries=time_retries, other_options=other_options)
+
+        print(f'Running output for: {output_file_name}. Started {time.strftime("%H:%M:%S %d.%m.")}')
+        with open(f'{output_dir}/{output_file_name}', "w") as output_file:
+            for solution in it:
+                # FIXME: Parse solution dict
+                write_solution(output_file, solution)
+                output_file.flush()
